+++ CHANGES.md
@@ -51,6 +51,7 @@ Changes in 12.5
 - JS remove inline code from program in Scheduler.php & Scheduler.js
 - JS remove inline code from program in AccessLog.php & AccessLog.js
 - JS remove inline code from program in Configuration.php & Configuration.js
+- JS remove inline code from program in MarkingPeriods.php & MarkingPeriods.js
 
 Changes in 12.4.3
 -----------------